requirements.txt
fastapi
uvicorn[standard]
motor
PyJWT

